Layout resources representing the activities.
